# backend-gin/internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/auth/signin": {
            "post": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User authentication",
                "parameters": [
                    {
                        "minLength": 8,
                        "type": "string",
                        "example": "very_secret",
                        "description": "The user password, 8 characters at least",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "mslimbeji@gmail.com",
                        "description": "The user email (We use username here because of OAuth spec)",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.EncodedToken"
                        }
                    }
                }
            }
        },
        "/api/auth/signup": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User registration",
                "parameters": [
                    {
                        "type": "string",
                        "example": "mslimbeji@gmail.com",
                        "description": "The user email",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "minLength": 2,
                        "type": "string",
                        "example": "Slim Beji",
                        "description": "The user name, two characters at least",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "minLength": 8,
                        "type": "string",
                        "example": "very_secret",
                        "description": "The user password, 8 characters at least",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "User's profile image (JPEG)",
                        "name": "image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.EncodedToken"
                        }
                    }
                }
            }
        },
...
        "/api/hello-world/admin": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": [
                            "admin"
                        ]
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hello World"
                ],
                "summary": "Hello World Endpoint for admins only",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.HelloAdminResponse"
                        }
                    }
                }
            }
        },
        "/api/hello-world/user": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": [
                            "admin"
                        ]
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hello World"
                ],
                "summary": "Hello World Endpoint for authenticated users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.HelloUserResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "routes.HelloAdminResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Hello Admin Slim"
                }
            }
        },
...
        "routes.HelloUserResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Hello Slim"
                }
            }
        },
...
        "schemas.EncodedToken": {
            "type": "object",
            "properties": {
                "access_token": {
                    "description": "A generated web token. The 'Bearer ' prefix needs to be added for authentication",
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2ODIyNDVhOWY2YTU5ZjVlNjM2Y2NmYjEiLCJlbWFpbCI6ImJlamkuc2xpbUB5YWhvby5mciIsImlhdCI6MTc0NzMzNjUxMCwiZXhwIjoxNzQ3MzQwMTEwfQ.C4DCJKvGWhpHClpqmxHyxKLPYDOZDUlr-LA_2IflTXM"
                },
                "email": {
                    "description": "The user email",
                    "type": "string",
                    "example": "mslimbeji@gmail.com"
                },
                "expires_in": {
                    "description": "The UNIX timestamp the token expires at",
                    "type": "integer",
                    "example": 1751879562
                },
                "token_type": {
                    "description": "The type of token. Only 'bearer' is supported.",
                    "type": "string",
                    "enum": [
                        "bearer"
                    ],
                    "example": "bearer"
                },
                "userId": {
                    "description": "The user ID, 24 characters",
                    "type": "integer",
                    "example": 123456789
                }
            }
        },
...
    },
    "securityDefinitions": {
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/api/auth/signin"
        }
    },
...
}